requirements.txt
asgiref==3.7.2
attrs==23.2.0
az-iranian-bank-gateways==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
django-js-asset==2.2.0
django-mptt==0.16.0
django-phonenumber-field==7.3.0
idna==3.6
isodate==0.6.1
kavenegar==1.1.2
lxml==5.1.0
phonenumbers==8.13.32
pillow==10.2.0
platformdirs==4.2.0
pycryptodome==3.20.0
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
zeep==4.2.1
